Show CM performance (and incidentally "My Own Reports) to everyone in the Reports Center
