+++ src/components/NavBar/NavBar.tsx
@@ -290,12 +290,11 @@ export function NavBar(): JSX.Element {
                         {_("Rating Calculator")}
                     </Link>
 
-                    {(user.is_moderator || !!user.moderator_powers) && (
-                        <Link className="admin-link" to="/reports-center">
-                            <i className="fa fa-exclamation-triangle"></i>
-                            {_("Reports Center")}
-                        </Link>
-                    )}
+                    <Link className="admin-link" to="/reports-center">
+                        <i className="fa fa-exclamation-triangle"></i>
+                        {_("Reports Center")}
+                    </Link>
+
                     {user.is_moderator && (
                         <Link className="admin-link" to="/moderator">
                             <i className="fa fa-gavel"></i>
